# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/admin/orders": {
      "get": {
        "tags": [
          "Admin Panel"
        ],
        "summary": "Fetches data for orders page",
        "description": "You need to execute /auth/login as admin first",
        "produces": [
          "json"
        ],
        "responses": {
          "401": {
            "description": "Unauthorized: No token provided or invalid token"
          },
          "403": {
            "description": "Unauthorized: Only admins have access"
          },
          "500": {
            "description": "Internal Server Error: Error fetching data for orders page"
          }
        }
      }
    },
...
    "/orders/": {
      "get": {
        "tags": [
          "Checkout and Orders"
        ],
        "summary": "View all orders of all users",
        "description": "",
        "responses": {
          "200": {
            "description": "OK: Successful operation"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error: An error occurred while trying to view orders"
          }
        }
      }
    },
    "/orders/statuses": {
      "get": {
        "tags": [
          "Checkout and Orders"
        ],
        "summary": "View status alternatives",
        "description": "",
        "responses": {
          "200": {
            "description": "OK: Successful operation"
          },
          "404": {
            "description": "Not Found: No statuses found"
          },
          "500": {
            "description": "Internal Server Error: An error occurred while trying to view statuses"
          }
        }
      }
    },
    "/orders/{orderId}": {
      "put": {
        "tags": [
          "Checkout and Orders"
        ],
        "summary": "Change order statuses by orderId",
        "description": "You need to execute /auth/login as admin first",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateOrderStatus"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created: Status updated successfully"
          },
          "400": {
            "description": "Bad Request: Invalid properties"
          },
          "401": {
            "description": "Unauthorized: No token provided or invalid token"
          },
          "403": {
            "description": "Unauthorized: Only admin have access"
          },
          "404": {
            "description": "Not Found: Order or Status not found"
          },
          "500": {
            "description": "Internal Server Error: An error occurred while trying to update order status"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "UpdateOrderStatus": {
      "type": "object",
      "properties": {
        "statusName": {
          "type": "string",
          "example": "Completed"
        }
      }
    },
...
  }
}